+++ go.mod
@@ -2,7 +2,8 @@ module github.com/arajski/socket-chat
 
 go 1.21
 
-require internal/server v1.0.0
-replace internal/server => ./internal/server
-require internal/client v1.0.0
-replace internal/client => ./internal/client
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/cobra v1.7.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
