--- internal/client/go.mod
@@ -1,3 +1,0 @@
-module client
-
-go 1.21
