--- internal/server/go.mod
@@ -1,3 +1,0 @@
-module server
-
-go 1.21
