+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/arajski/socket-chat/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
